+++ app/src/main/java/com/skydoves/disneycompose/ui/details/DetailViewModel.kt
@@ -17,7 +17,6 @@
 package com.skydoves.disneycompose.ui.details
 
 import androidx.lifecycle.ViewModel
-import com.skydoves.disneycompose.repository.DetailRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.flatMapLatest
