+++ app/src/main/java/com/skydoves/disneycompose/ui/main/MainViewModel.kt
@@ -22,7 +22,6 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.skydoves.disneycompose.model.Poster
-import com.skydoves.disneycompose.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import timber.log.Timber
